\name{getGenerations}
\alias{getGenerations}
\alias{getGenerations-methods}
\alias{getGenerations,proliferationFittingData-method}

\title{ Get percentage of cells for generation in a flowFit model }

\description{
getGenerations:  get percentage of cells for generation in a flowFit model from an object of class \code{\link{proliferationFittingData}} generated by the \code{\link{proliferationFitting}} function.
}

\usage{
getGenerations(object)
}

\arguments{
  \item{object}{
    An object of class \code{proliferationFittingData}
  }
}

\value{
return a \code{list} object
}

\details{
  This function return a \code{list}.
  In order to get the percentage of cells for generation as vector you can use th \code{slot} \code{generations} of the \code{\link{proliferationFittingData}} (see also examples).
}

\author{ Davide Rambaldi }

\seealso{
  \code{\link{proliferationFitting}} and \code{\link{proliferationFittingData}}
}

\examples{
  if(require(flowFitExampleData)){
    data(QuahAndParish)
    parent.fitting.cfse <- parentFitting(QuahAndParish[[1]], "<FITC-A>")
    fitting.cfse <- proliferationFitting(QuahAndParish[[2]], "<FITC-A>", 
                                         parent.fitting.cfse@parentPeakPosition, 
                                         parent.fitting.cfse@parentPeakSize)
    generationList <- getGenerations(fitting.cfse)

    # to extract a vector of percentage of cells for generation you can use:
    fitting.cfse@generations
  }
}