\name{plot-methods}
\docType{methods}
\alias{plot}
\alias{plot-methods}
\alias{plot,proliferationFittingData,ANY-method}
\alias{plot,proliferationFittingData,character-method}
\alias{plot,proliferationFittingData,missing-method}
\alias{plot,parentFittingData,ANY-method}
\alias{plot,parentFittingData,character-method}
\alias{plot,parentFittingData,missing-method}

\title{ Very basic plotting of flowFit objects: \code{\link{proliferationFittingData}} and \code{\link{parentFittingData}} }


\description{ A basic method to plot \code{\link{proliferationFittingData}} and \code{\link{parentFittingData}}
  objects. See below for details.  }

\section{Methods}{
  \describe{
    \item{x = "proliferationFittingData", y = "missing"}{ Multiple plots of generations fitting data (data generated by the function \link{proliferationFitting}) }
    \item{x = "parentFittingData", y= "missing")}{ Single plot of a parent population fitting (data generated by the function \link{parentFitting})}
  }
}

\details{
  Basic plots for \code{flowFit} objects.

  Supported arguments for \link{parentFittingData}:

  \enumerate{
    \item \code{main}: an overall title for the plot, see \link{title}.
    \item \code{xlab}: a title for the x axis, see \link{title}.
    \item \code{ylab}: a title for the y axis, see \link{title}.
    \item \code{legend}: show/hide messages.
    \item \code{logScale}: put a log scale on the x axis.
  }

  Supported arguments for \link{proliferationFittingData}:

  \enumerate{
    \item \code{which}: which plots I should show? ["all" or 1:5]
    \item \code{main}: an overall title prefix for the plots.
    \item \code{xlab}: a title for the x axis, see \link{title}.
    \item \code{ylab}: a title for the y axis, see \link{title}.
    \item \code{legend}: show/hide messages.
    \item \code{logScale}: put a log scale on the x axis.
    \item \code{drawGrid}: put some dashed lines at the \link{generationsDistance} expected positions.
  }

}


\author{ Davide Rambaldi }

\seealso{

  \code{\link{proliferationFitting}},
  \code{\link{parentFitting}}
}

\keyword{methods}


